Add all(if not missed) models

+++ src/commonMain/kotlin/methods.kt
@@ -7,9 +7,9 @@ sealed class Method
 
 @Serializable
 class GetUpdates private constructor(
-    val offset: Long?,
-    val limit: Long?,
-    val timeout: Long?,
+    val offset: Int?,
+    val limit: Int?,
+    val timeout: Int?,
     val allowed_updates: List<String>?
 ) : Method() {
 
@@ -18,9 +18,9 @@ class GetUpdates private constructor(
     }
 
     class GetUpdatesBuilder {
-        var offset: Long? = null
-        var limit: Long? = null
-        var timeout: Long? = null
+        var offset: Int? = null
+        var limit: Int? = null
+        var timeout: Int? = null
         var allowed_updates: List<String>? = null
 
         fun build() = GetUpdates(offset = offset, limit = limit, timeout = timeout, allowed_updates = allowed_updates)
@@ -30,12 +30,12 @@ class GetUpdates private constructor(
 
 @Serializable
 class SendMessage private constructor(
-    val chat_id: Long,
+    val chat_id: Int,
     val text: String,
     val parse_mode: String?,
     val disable_web_page_preview: Boolean?,
     val disable_notification: Boolean?,
-    val reply_to_message_id: Long?
+    val reply_to_message_id: Int?
 ) : Method() {
 
     companion object {
@@ -46,9 +46,9 @@ class SendMessage private constructor(
         var parse_mode: String? = "HTML"
         var disable_web_page_preview: Boolean? = null
         var disable_notification: Boolean? = null
-        var reply_to_message_id: Long? = null
+        var reply_to_message_id: Int? = null
 
-        fun build(chat_id: Long, text: String) =
+        fun build(chat_id: Int, text: String) =
             SendMessage(
                 chat_id = chat_id,
                 text = text,
