+++ src/commonMain/kotlin/models/Update.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Update(
-    val update_id: Long,
+    val update_id: Int,
     val message: Message? = null
 )
